# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users/me/preferences": {
      "patch": {
        "operationId": "UsersController_updateMyPreferences_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/roles": {
      "patch": {
        "operationId": "UsersController_assignRoles_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/status": {
      "patch": {
        "operationId": "UsersController_updateStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/contact-messages": {
      "post": {
        "operationId": "ContactMessagesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "contact-messages"
        ]
      },
      "get": {
        "operationId": "ContactMessagesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "contact-messages"
        ]
      }
    },
    "/api/v1/contact-messages/{id}": {
      "get": {
        "operationId": "ContactMessagesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "contact-messages"
        ]
      },
      "delete": {
        "operationId": "ContactMessagesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "contact-messages"
        ]
      }
    },
    "/api/v1/contact-messages/{id}/reply": {
      "patch": {
        "operationId": "ContactMessagesController_reply_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyToContactMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "contact-messages"
        ]
      }
    }
  },
  "info": {
    "title": "UAEAF Backend API",
    "description": "UAE Athletics Federation platform API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LocalizedTextDto": {
        "type": "object",
        "properties": {
          "en": {
            "type": "string",
            "description": "English text."
          },
          "ar": {
            "type": "string",
            "description": "Arabic text."
          }
        },
        "required": [
          "en",
          "ar"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "description": "Full name, recorded in both Arabic and English.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "email": {
            "type": "string",
            "description": "Login email, must be unique."
          },
          "password": {
            "type": "string",
            "description": "Plaintext password — hashed before storage, never persisted as-is."
          },
          "roleIds": {
            "description": "Roles the account starts with. Validated against the live roles collection; an unknown or archived id rejects the whole request. Omit for an account with no access.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "personId": {
            "type": "string",
            "description": "Optional link to the federationPersonnel record this account belongs to. Verified to exist before the account is written.",
            "nullable": true
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
...
      "UpdatePreferencesDto": {
        "type": "object",
        "properties": {
          "preferredLanguage": {
            "type": "string",
            "description": "Preferred dashboard language, or null to follow the request locale.",
            "enum": [
              "ar",
              "en"
            ],
            "nullable": true
          },
          "preferredTheme": {
            "type": "string",
            "description": "Preferred colour theme, or null to follow the OS setting.",
            "enum": [
              "light",
              "dark"
            ],
            "nullable": true
          }
        }
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "email": {
            "type": "string"
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "personId": {
            "type": "object",
            "nullable": true
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "Active",
              "Suspended",
              "Deactivated"
            ]
          },
          "lastLogin": {
            "type": "object",
            "nullable": true
          },
          "photoId": {
            "type": "object",
            "nullable": true
          },
          "preferredLanguage": {
            "type": "string",
            "enum": [
              "ar",
              "en"
            ],
            "nullable": true
          },
          "preferredTheme": {
            "type": "string",
            "enum": [
              "light",
              "dark"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "roleIds",
          "accountStatus"
        ]
      },
      "AssignRolesDto": {
        "type": "object",
        "properties": {
          "roleIds": {
            "description": "Full replacement list of role ids for this user.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roleIds"
        ]
      },
      "UpdateAccountStatusDto": {
        "type": "object",
        "properties": {
          "accountStatus": {
            "type": "string",
            "description": "The account state to move to. Anything other than Active also ends every live session for that account.",
            "enum": [
              "Active",
              "Suspended",
              "Deactivated"
            ]
          }
        },
        "required": [
          "accountStatus"
        ]
      },
...
      "CreateContactMessageDto": {
        "type": "object",
        "properties": {
          "messageType": {
            "type": "string",
            "enum": [
              "Complaint",
              "Suggestion",
              "Inquiry",
              "General"
            ]
          },
          "senderName": {
            "type": "string"
          },
          "senderEmail": {
            "type": "string"
          },
          "senderPhone": {
            "type": "string"
          },
          "messageBody": {
            "type": "string"
          }
        },
        "required": [
          "messageType",
          "senderName",
          "senderEmail",
          "messageBody"
        ]
      },
      "ReplyToContactMessageDto": {
        "type": "object",
        "properties": {
          "replyBody": {
            "type": "string",
            "description": "The reply text. Recording only — the system does not send it."
          },
          "replyChannel": {
            "type": "string",
            "enum": [
              "Email",
              "Phone"
            ]
          }
        },
        "required": [
          "replyBody",
          "replyChannel"
        ]
      }
    }
  }
}